requirements.txt
sqlalchemy
alembic
pydantic
email-validator
passlib
bcrypt==4.0.1
passlib[bcrypt]
python-jose
pg8000

boto3

pytest
pytest-watch
polyfactory
tbvaccine

wheel
setuptools

pylsp-mypy
types-passlib
types-python-jose
colored-traceback

snakecase
colorama

flask

locust

injector
pyright
